Minor link/wording changes
We used your tutorials in a workshop yesterday at Princeton CDH, and a less technical person at my table seemed to be struggling a bit with downloading the raw CSV vs the html representation, so I added a little more help text and a link directly to the CSV file. (I couldn't figure out a way to do a relative link to the raw CSV, but if you're not changing it I think this should be fine).  Also, when I went back into the document to help him download the file the repeated 'here' links were hard to differentiate.